fix(e2e): split smoke suite — reuse setup-owner storageState
Fixes the post-merge CI flake on busy days (4+ PRs merging within an hour
hammered Supabase Auth's ~45-sign-ins/minute limit). The smoke suite
previously called loginAsOwner() through the UI inside 4 of 5 tests, so
each CI run cost 5 UI logins.

Architecture: split the suite so the existing setup-owner project's
storageState gets reused.

- tests/e2e/tests/smoke/critical-paths.smoke.spec.ts: trimmed to the one
  test that genuinely exercises the login UI (Owner can login) + the
  environment sanity checks. Still runs un-authenticated under the
  smoke project.
- tests/e2e/tests/smoke/authenticated-flows.smoke.spec.ts: new file with
  the 4 post-login tests (Dashboard / Properties / Navigation / No
  console errors). Each starts with page.goto() directly — no
  loginAsOwner() call. Storage state injected by playwright project.
- tests/e2e/playwright.config.ts: added smoke-authenticated project with
  dependencies: ['setup-owner'] and storageState: OWNER_AUTH_FILE.
  Original smoke project narrowed via testMatch.
- .github/workflows/ci-cd.yml: e2e step now invokes --project=smoke
  --project=smoke-authenticated --project=public.

Net auth-call cost per CI run drops from 5 UI logins to 1 Supabase Auth
API call (the setup-owner script, which is ~500ms and rate-limit-safe).
The "P0 Dashboard loads for owner" flake captured in MEMORY.md no longer
applies.

99679 unit tests pass; typecheck clean; lint clean;
bunx playwright test --list shows both projects discover correctly.

[hudsor01]
